+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log
 
-All notable changes to the "CSV-i18n" extension will be documented in this file.
+All notable changes to the "JTool" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
 - Initial release
+- Added icons for "CSV and JSON Operations" and "Compare HTML Structure" tree items
+- Added icons for children items under "CSV and JSON Operations" and "Compare HTML Structure"